+++ src/exchange.ts
@@ -788,6 +788,16 @@ export class Exchange {
     }
     this._programSubscriptionIds = [];
   }
+
+  public async initializeReferrerAccount(referrer: PublicKey) {
+    let tx = new Transaction().add(
+      await instructions.initializeReferrerAccountIx(
+        referrer,
+        this._provider.wallet.publicKey
+      )
+    );
+    await utils.processTransaction(this._provider, tx);
+  }
 }
 
 // Exchange singleton.