+++ CHANGELOG.md
@@ -7,7 +7,12 @@ Version changes are pinned to SDK releases.
 
 ## [0.15.0]
 
+## [0.14.4]
+
 - client: Add extra safety around client state updates. ([#129](https://github.com/zetamarkets/sdk/pull/129))
+- client: Add referral logic for client. ([#130](https://github.com/zetamarkets/sdk/pull/130))
+- program: Add instruction logic for referrals. ([#130](https://github.com/zetamarkets/sdk/pull/130))
+- events: Cancel events. ([#131](https://github.com/zetamarkets/sdk/pull/131))
 
 ## [0.14.3]